+++ Sources/MapboxSearch/InternalAPI/Telemetry/EventsServiceProtocol.swift
@@ -1,0 +1,8 @@
+import Foundation
+import MapboxCommon_Private
+
+protocol EventsServiceProtocol {
+    func sendEvent(for event: Event, callback: EventsServiceResponseCallback?)
+}
+
+extension EventsService: EventsServiceProtocol { }